chore: refactor create event types form
