+++ packages/features/eventtypes/components/CreateEventTypeForm.tsx
@@ -1,0 +1,136 @@
+import type { ReactNode } from "react";
+import { useState } from "react";
+import type { UseFormReturn } from "react-hook-form";
+
+import { useIsPlatform } from "@calcom/atoms/monorepo";
+import type { CreateEventTypeFormValues } from "@calcom/lib/hooks/useCreateEventType";
+import { useLocale } from "@calcom/lib/hooks/useLocale";
+import { md } from "@calcom/lib/markdownIt";
+import slugify from "@calcom/lib/slugify";
+import turndown from "@calcom/lib/turndownService";
+import { Editor, Form, TextAreaField, TextField, Tooltip } from "@calcom/ui";
+
+export default function CreateEventTypeForm({
+  form,
+  isManagedEventType,
+  handleSubmit,
+  pageSlug,
+  isPending,
+  urlPrefix,
+  SubmitButton,
+}: {
+  form: UseFormReturn<CreateEventTypeFormValues>;
+  isManagedEventType: boolean;
+  handleSubmit: (values: CreateEventTypeFormValues) => void;
+  pageSlug?: string;
+  isPending: boolean;
+  urlPrefix?: string;
+  SubmitButton: (isPending: boolean) => ReactNode;
+}) {
+  const isPlatform = useIsPlatform();
+  const { t } = useLocale();
+  const [firstRender, setFirstRender] = useState(true);
+
+  const { register } = form;
+  return (
+    <Form
+      form={form}
+      handleSubmit={(values) => {
+        handleSubmit(values);
+      }}>
+      <div className="mt-3 space-y-6 pb-11">
+        <TextField
+          label={t("title")}
+          placeholder={t("quick_chat")}
+          data-testid="event-type-quick-chat"
+          {...register("title")}
+          onChange={(e) => {
+            form.setValue("title", e?.target.value);
+            if (form.formState.touchedFields["slug"] === undefined) {
+              form.setValue("slug", slugify(e?.target.value));
+            }
+          }}
+        />
+
+        {urlPrefix && urlPrefix.length >= 21 ? (
+          <div>
+            <TextField
+              label={`${t("url")}: ${urlPrefix}`}
+              required
+              addOnLeading={
+                !isPlatform ? (
+                  <Tooltip content={!isManagedEventType ? pageSlug : t("username_placeholder")}>
+                    <span className="max-w-24 md:max-w-56">
+                      /{!isManagedEventType ? pageSlug : t("username_placeholder")}/
+                    </span>
+                  </Tooltip>
+                ) : undefined
+              }
+              {...register("slug")}
+              onChange={(e) => {
+                form.setValue("slug", slugify(e?.target.value), { shouldTouch: true });
+              }}
+            />
+
+            {isManagedEventType && (
+              <p className="mt-2 text-sm text-gray-600">{t("managed_event_url_clarification")}</p>
+            )}
+          </div>
+        ) : (
+          <div>
+            <TextField
+              label={t("url")}
+              required
+              addOnLeading={
+                !isPlatform ? (
+                  <Tooltip
+                    content={`${urlPrefix}/${!isManagedEventType ? pageSlug : t("username_placeholder")}/`}>
+                    <span className="max-w-24 md:max-w-56">
+                      {urlPrefix}/{!isManagedEventType ? pageSlug : t("username_placeholder")}/
+                    </span>
+                  </Tooltip>
+                ) : undefined
+              }
+              {...register("slug")}
+              onChange={(e) => {
+                form.setValue("slug", slugify(e?.target.value), { shouldTouch: true });
+              }}
+            />
+            {isManagedEventType && (
+              <p className="mt-2 text-sm text-gray-600">{t("managed_event_url_clarification")}</p>
+            )}
+          </div>
+        )}
+        <>
+          {isPlatform ? (
+            <TextAreaField {...register("description")} placeholder={t("quick_video_meeting")} />
+          ) : (
+            <Editor
+              getText={() => md.render(form.getValues("description") || "")}
+              setText={(value: string) => form.setValue("description", turndown(value))}
+              excludedToolbarItems={["blockType", "link"]}
+              placeholder={t("quick_video_meeting")}
+              firstRender={firstRender}
+              setFirstRender={setFirstRender}
+              maxHeight="200px"
+            />
+          )}
+
+          <div className="relative">
+            <TextField
+              type="number"
+              required
+              min="10"
+              placeholder="15"
+              label={t("duration")}
+              className="pr-4"
+              {...register("length", { valueAsNumber: true })}
+              addOnSuffix={t("minutes")}
+            />
+          </div>
+        </>
+      </div>
+      {SubmitButton(isPending)}
+    </Form>
+  );
+}
