+++ packages/lib/hooks/useCreateEventType.ts
@@ -11,13 +11,15 @@ import { unlockedManagedEventTypeProps } from "@calcom/prisma/zod-utils";
 import { createEventTypeInput } from "@calcom/prisma/zod/custom/eventtype";
 import { trpc } from "@calcom/trpc/react";
 
+export type CreateEventTypeFormValues = z.infer<typeof createEventTypeInput>;
+
 export const useCreateEventType = (
   onSuccessMutation: (eventType: EventType) => void,
   onErrorMutation: (message: string) => void
 ) => {
   const utils = trpc.useUtils();
   const { t } = useLocale();
-  const form = useForm<z.infer<typeof createEventTypeInput>>({
+  const form = useForm<CreateEventTypeFormValues>({
     defaultValues: {
       length: 15,
     },
